Moved dbinit
